Moved examples in src module.
