+++ src/examples/shaders/fragment_shader_code_red.glsl
