+++ src/examples/shaders/textured_fragment_shader_code_red.glsl
