+++ src/examples/shaders/textured_vertex_shader_code.glsl
