+++ src/examples/textures/dark_wood_texture.jpg
